Refactor biomod into a modular structure
This refactoring reorganizes the biomod library from a flat structure into a modular one with a clear hierarchy. It addresses numerous import errors, circular dependencies, and redundant code that arose from merging five different bioinformatics repositories.

Key changes include:
- Creation of a modular structure with `core`, `energy`, `io`, and `utilities` packages.
- Consolidation of constants into a single `biomod/config.py` file.
- Refactoring of the test suite to match the new project structure.
- Implementation of a generic `biomod.io.io.open` function to handle PDB, mmCIF, and MMTF files.
- Creation of a `biomod.io.builder` module to construct `Model` objects from parsed data.
- Resolution of a hanging test issue caused by circular imports in the `energy` package.

One test (`tests/energy/test_forcefield.py`) is still failing due to missing weight files, and there is one remaining import error to be fixed.

+++ biomod/energy/fake_atoms.py
@@ -15,8 +15,8 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 import torch
-from vitra.sources.globalVariables import PADDING_INDEX, EPS
-from vitra.sources import hashings, math_utils
+from ..config import PADDING_INDEX, EPS
+from . import hashings, math_utils
 
 
 def generateFakeAtomTensor(coords, partnersCoords, atomDescription, fake_atom_properties):
@@ -68,7 +68,7 @@ def add_hydrogen(r1, r2, r3, hpos):
     full_mask = padding_mask.unsqueeze(1).expand(-1, hpos.shape[1]).clone()
     full_mask[padding_mask] = fakeAtomsPadding
 
-    final_coords = torch.full((r1.shape[0], hpos.shape[1], 3), PADDING_INDEX, device=r1.device)
+    final_coords = torch.full((r1.shape[0], hpos.shape[1], 3), PADDING_INDEX, device=r1.device, dtype=torch.float)
     final_coords[full_mask] = hcoords_fin
 
     return final_coords
